--- android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/common/LocalDateConverters.kt
@@ -1,10 +1,0 @@
-package com.woowacourse.staccato.data.common
-
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-
-const val DATETIME_FORMAT_PATTERN = "yyyy-MM-dd"
-
-fun String.toLocalDate(): LocalDate {
-    return LocalDate.parse(this, DateTimeFormatter.ofPattern(DATETIME_FORMAT_PATTERN))
-}
