+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/dto/mapper/TimelineMapper.kt
@@ -1,13 +1,10 @@
 package com.woowacourse.staccato.data.dto.mapper
 
-import com.woowacourse.staccato.data.common.toLocalDate
-import com.woowacourse.staccato.data.dto.MemberDto
-import com.woowacourse.staccato.data.dto.MembersDto
 import com.woowacourse.staccato.data.dto.timeline.TimelineResponse
 import com.woowacourse.staccato.data.dto.timeline.TimelineTravelDto
-import com.woowacourse.staccato.domain.model.Member
 import com.woowacourse.staccato.domain.model.Timeline
 import com.woowacourse.staccato.domain.model.Travel
+import java.time.LocalDate
 
 fun TimelineResponse.toDomain(): Timeline {
     val travels =
@@ -22,22 +19,10 @@ fun TimelineTravelDto.toDomain(): Travel {
         travelId = travelId,
         travelThumbnail = travelThumbnail,
         travelTitle = travelTitle,
-        startAt = startAt.toLocalDate(),
-        endAt = endAt.toLocalDate(),
+        startAt = LocalDate.parse(startAt),
+        endAt = LocalDate.parse(endAt),
         description = description,
         mates = mates.toDomain(),
         visits = emptyList(),
-    )
-}
-
-fun MembersDto.toDomain(): List<Member> {
-    return members.map { it.toDomain() }
-}
-
-fun MemberDto.toDomain(): Member {
-    return Member(
-        memberId = memberId,
-        nickName = nickName,
-        memberImage = memberImage,
     )
 }
