+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/dto/member/MemberDto.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.data.dto
+package com.woowacourse.staccato.data.dto.member
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
