+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/dto/travel/TravelResponse.kt
@@ -1,6 +1,6 @@
 package com.woowacourse.staccato.data.dto.travel
 
-import com.woowacourse.staccato.data.dto.MemberDto
+import com.woowacourse.staccato.data.dto.member.MemberDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
