+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/timeline/TimelineRemoteDataSource.kt
@@ -1,7 +1,6 @@
-package com.woowacourse.staccato.data.datasource
+package com.woowacourse.staccato.data.timeline
 
 import com.woowacourse.staccato.data.StaccatoClient
-import com.woowacourse.staccato.data.apiservice.TimeLineApiService
 import com.woowacourse.staccato.data.dto.timeline.TimelineResponse
 import org.json.JSONObject
