+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/visit/VisitDefaultRepository.kt
@@ -1,9 +1,8 @@
-package com.woowacourse.staccato.data.repository
+package com.woowacourse.staccato.data.visit
 
 import com.woowacourse.staccato.data.dto.mapper.toDomain
 import com.woowacourse.staccato.data.dto.visit.VisitCreationRequest
 import com.woowacourse.staccato.data.dto.visit.VisitUpdateRequest
-import com.woowacourse.staccato.data.visit.RemoteVisitDataSource
 import com.woowacourse.staccato.domain.model.Visit
 import com.woowacourse.staccato.domain.repository.VisitRepository
 import retrofit2.Response
