+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/visit/viewmodel/VisitViewModelFactory.kt
@@ -3,8 +3,8 @@ package com.woowacourse.staccato.presentation.visit.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.woowacourse.staccato.data.StaccatoClient.visitApiService
-import com.woowacourse.staccato.data.repository.VisitDefaultRepository
 import com.woowacourse.staccato.data.visit.RemoteVisitDataSource
+import com.woowacourse.staccato.data.visit.VisitDefaultRepository
 
 class VisitViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
